DietController

+++ EngineerApp.API/EngineerApp.API/Controllers/TrainingController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Engineer.Models.BindingModel.Training;
 using Engineer.Models.BindingModel.Training.Edit;
@@ -12,7 +13,7 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace EngineerApp.API.Controllers
 {
-    [AllowAnonymous]
+    [Authorize]
     public class TrainingController : BaseController
     {
         private readonly ITrainingService _trainingService;
@@ -22,9 +23,12 @@ namespace EngineerApp.API.Controllers
             _trainingService = trainingService;
         }
 
+        [Authorize(Policy = "RequireTrainerRole")]
         [HttpPost]
         public async Task<IActionResult> AddTraining(int idUser, int idTrainer, [FromBody]TrainingPlanBindingModel model)
         {
+            if (idTrainer != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                return Unauthorized();
 
             var result = await _trainingService.InsertTraining(idUser, idTrainer, model);
 
@@ -34,9 +38,13 @@ namespace EngineerApp.API.Controllers
             return Ok(result);
         }
 
+        [Authorize(Policy = "RequireTrainerRole")]
         [HttpDelete("{trainingId}")]
-        public async Task<IActionResult> DeleteTraining(int trainingId)
+        public async Task<IActionResult> DeleteTraining(int trainingId, int trainerId)
         {
+            if (trainerId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                return Unauthorized();
+
             var result = await _trainingService.DeleteTraining(trainingId);
 
             if (result.ErrorOccurred)
@@ -65,9 +73,12 @@ namespace EngineerApp.API.Controllers
             return Ok(result);
         }
 
-        [HttpGet("TrainingUsers/{userId}")]
+        [HttpGet("Users/{userId}")]
         public IActionResult GetTrainingUsers(int userId)
         {
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                return Unauthorized();
+
             var result = _trainingService.GetTrainingUsers(userId);
             if (result.ErrorOccurred)
                 return BadRequest(result);
@@ -75,9 +86,13 @@ namespace EngineerApp.API.Controllers
             return Ok(result);
         }
 
-        [HttpPut("Training/{trainingId}")]
-        public async Task<IActionResult> EditTraining(int trainingId, EditTrainingPlanBindingModel model)
+        [Authorize(Policy = "RequireTrainerRole")]
+        [HttpPut("Training")]
+        public async Task<IActionResult> EditTraining(int trainingId, int trainerId, EditTrainingPlanBindingModel model)
         {
+            if (trainerId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                return Unauthorized();
+
             var result = await _trainingService.EditTraining(trainingId, model);
             if (result.ErrorOccurred)
             {
@@ -87,9 +102,13 @@ namespace EngineerApp.API.Controllers
             return Ok(result);
         }
 
-        [HttpPut("day/{dayId}")]
-        public async Task<IActionResult> EditDay(int dayId, EditTrainingDayBindingModel model)
+        [Authorize(Policy = "RequireTrainerRole")]
+        [HttpPut("day")]
+        public async Task<IActionResult> EditDay(int dayId, int trainerId, EditTrainingDayBindingModel model)
         {
+            if (trainerId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                return Unauthorized();
+
             var result = await _trainingService.EditDay(dayId, model);
             if (result.ErrorOccurred)
             {
@@ -99,9 +118,13 @@ namespace EngineerApp.API.Controllers
             return Ok(result);
         }
 
-        [HttpPut("exercise/{exerciseId}")]
-        public async Task<IActionResult> EditExercise(int exerciseId, EditExerciseTrainingBindingModel model)
+        [Authorize(Policy = "RequireTrainerRole")]
+        [HttpPut("exercise")]
+        public async Task<IActionResult> EditExercise(int exerciseId, int trainerId, EditExerciseTrainingBindingModel model)
         {
+            if (trainerId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                return Unauthorized();
+
             var result = await _trainingService.EditExercise(exerciseId, model);
             if (result.ErrorOccurred)
             {
@@ -111,9 +134,13 @@ namespace EngineerApp.API.Controllers
             return Ok(result);
         }
 
-        [HttpPut("serie/{serieId}")]
-        public async Task<IActionResult> EditSerie(int serieId, EditSerieBindingModel model)
+        [Authorize(Policy = "RequireTrainerRole")]
+        [HttpPut("serie")]
+        public async Task<IActionResult> EditSerie(int serieId, int trainerId, EditSerieBindingModel model)
         {
+            if (trainerId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
+                return Unauthorized();
+
             var result = await _trainingService.EditSerie(serieId, model);
             if (result.ErrorOccurred)
             {
